[sui-execution] First step towards allowing versioning move-vm-types
