+++ crates/sui-types/src/layout_resolver.rs
@@ -1,23 +1,18 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::error::{ExecutionError, SuiError};
+use crate::error::SuiError;
 use move_bytecode_utils::{layout::TypeLayoutBuilder, module_cache::GetModule};
 use move_core_types::{
     annotated_value as A,
     language_storage::{StructTag, TypeTag},
 };
-use move_vm_types::loaded_data::runtime_types::Type;
 
 pub trait LayoutResolver {
     fn get_annotated_layout(
         &mut self,
         struct_tag: &StructTag,
     ) -> Result<A::MoveDatatypeLayout, SuiError>;
-}
-
-pub trait TypeTagResolver {
-    fn get_type_tag(&self, type_: &Type) -> Result<TypeTag, ExecutionError>;
 }
 
 pub fn get_layout_from_struct_tag(
