+++ crates/sui-types/src/lib.rs
@@ -48,7 +48,6 @@ pub mod event;
 pub mod executable_transaction;
 pub mod execution;
 pub mod execution_config_utils;
-pub mod execution_mode;
 pub mod execution_status;
 pub mod full_checkpoint_content;
 pub mod gas;
@@ -58,6 +57,7 @@ pub mod governance;
 pub mod id;
 pub mod in_memory_storage;
 pub mod inner_temporary_store;
+pub mod layout_resolver;
 pub mod message_envelope;
 pub mod messages_checkpoint;
 pub mod messages_consensus;
@@ -83,7 +83,6 @@ pub mod supported_protocol_versions;
 pub mod traffic_control;
 pub mod transaction;
 pub mod transfer;
-pub mod type_resolver;
 pub mod versioned;
 pub mod zk_login_authenticator;
 pub mod zk_login_util;
