+++ sui-execution/latest/sui-adapter/src/lib.rs
@@ -7,7 +7,10 @@ extern crate sui_types;
 pub mod adapter;
 pub mod error;
 pub mod execution_engine;
+pub mod execution_mode;
+pub mod execution_value;
 pub mod gas_charger;
 pub mod programmable_transactions;
 pub mod temporary_store;
 pub mod type_layout_resolver;
+pub mod type_resolver;
