+++ sui-execution/src/v1.rs
@@ -13,13 +13,12 @@ use sui_types::{
     digests::TransactionDigest,
     effects::TransactionEffects,
     error::{ExecutionError, SuiError, SuiResult},
-    execution::TypeLayoutStore,
-    execution_mode::{self, ExecutionResult},
+    execution::{ExecutionResult, TypeLayoutStore},
     gas::SuiGasStatus,
     inner_temporary_store::InnerTemporaryStore,
+    layout_resolver::LayoutResolver,
     metrics::{BytecodeVerifierMetrics, LimitsMetrics},
     transaction::{CheckedInputObjects, ProgrammableTransaction, TransactionKind},
-    type_resolver::LayoutResolver,
 };
 
 use move_bytecode_verifier_meter::Meter;
@@ -35,6 +34,7 @@ use sui_verifier_v1::meter::SuiVerifierMeter;
 
 use crate::executor;
 use crate::verifier;
+use sui_adapter_v1::execution_mode;
 
 pub(crate) struct Executor(Arc<MoveVM>);
