+++ sui-execution/v0/sui-adapter/src/programmable_transactions/linkage_view.rs
@@ -7,6 +7,7 @@ use std::{
     str::FromStr,
 };
 
+use crate::execution_value::SuiResolver;
 use move_core_types::{
     account_address::AccountAddress,
     identifier::{IdentStr, Identifier},
@@ -17,7 +18,6 @@ use sui_types::storage::{get_module, PackageObject};
 use sui_types::{
     base_types::ObjectID,
     error::{ExecutionError, SuiError, SuiResult},
-    execution::SuiResolver,
     move_package::{MovePackage, TypeOrigin, UpgradeInfo},
     storage::BackingPackageStore,
 };
