+++ sui-execution/v0/sui-adapter/src/temporary_store.rs
@@ -14,9 +14,9 @@ use sui_types::execution::{DynamicallyLoadedObjectMetadata, ExecutionResults, Sh
 use sui_types::execution_config_utils::to_binary_config;
 use sui_types::execution_status::ExecutionStatus;
 use sui_types::inner_temporary_store::InnerTemporaryStore;
+use sui_types::layout_resolver::LayoutResolver;
 use sui_types::storage::{BackingStore, DeleteKindWithOldVersion, DenyListResult, PackageObject};
 use sui_types::sui_system_state::{get_sui_system_state_wrapper, AdvanceEpochParams};
-use sui_types::type_resolver::LayoutResolver;
 use sui_types::{
     base_types::{
         ObjectDigest, ObjectID, ObjectRef, SequenceNumber, SuiAddress, TransactionDigest,
