+++ sui-execution/v1/sui-adapter/src/type_resolver.rs
@@ -1,0 +1,10 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+use move_core_types::language_storage::TypeTag;
+use move_vm_types::loaded_data::runtime_types::Type;
+use sui_types::error::ExecutionError;
+
+pub trait TypeTagResolver {
+    fn get_type_tag(&self, type_: &Type) -> Result<TypeTag, ExecutionError>;
+}
